+++ app/build.gradle
@@ -10,7 +10,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.1"
     defaultConfig {
-        applicationId "com.adriangl.pokeapi_mvvm"
+        applicationId "com.bq.arch_example"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
