+++ app/src/main/AndroidManifest.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.adriangl.pokeapi_mvvm">
+    package="com.bq.arch_example">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".App"
+        android:name="com.bq.arch_example.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name="com.adriangl.pokeapi_mvvm.ui.pokemonlist.PokemonListActivity">
+        <activity android:name=".ui.pokemonlist.PokemonListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
