--- app/src/main/kotlin/com/adriangl/pokeapi_mvvm/domain/moves/MoveActions.kt
@@ -1,10 +1,0 @@
-package com.adriangl.pokeapi_mvvm.domain.moves
-
-import com.adriangl.pokeapi_mvvm.network.MoveName
-import com.adriangl.pokeapi_mvvm.network.PokemonMove
-import mini.BaseAction
-import mini.Task
-
-data class LoadMovesAction(val moveNameList: List<MoveName>) : BaseAction()
-
-data class MoveLoadedAction(val moveName: MoveName, val pokemonMove: PokemonMove?, val task: Task) : BaseAction()
