+++ app/src/main/kotlin/com/bq/arch_example/domain/pokemon/PokemonActions.kt
@@ -1,0 +1,12 @@
+package com.bq.arch_example.domain.pokemon
+
+import com.bq.arch_example.network.Pokemon
+import mini.BaseAction
+import mini.Task
+
+
+class GetPokemonDetailsListAction : BaseAction()
+
+data class PokemonDetailsListLoadedAction(val pokemonList: List<Pokemon>?, val task: Task) : BaseAction()
+
+data class FilterPokemonListAction(val query: String) : BaseAction()
